Update Torch QR dispatch
